probleme cu stiva, probleme diverse de clasa a 9-a

+++ pbinfo/clasa9/AlgoritmiElementari/Maxime si minime/ProdusMaxime#4546/ProdusMaxime.cpp
@@ -25,7 +25,7 @@ int main()
     }
     else
     {
-        cout << max1 * max2;
+        cout << (long long)(max1 * max2);
     }
     return 0;
 }
